Regenerate transient key when changing locations
Previously, the transient key from the user location at the begining of the function was re-used later in the function, after the location had changed. This led to the events for one location being stored in the transient for a different location, and the wrong events would be displayed when the user loaded a fresh copy of the Dashboard.

Fixes #10