Rename `$events` to `$response_body` for accuracy and clarity
The variable is not just an array of events as its previous name would imply, it actually contains the body of the response to an HTTP request. Naming it `$events` is misleading, and results in awkward code like `$events['events']`.

Naming this accurately also helps to make it clear that the data we're caching, returning to calling functions, and eventually passing to JavaScript actually needs to be a combination of the location and the events, not just the events.