Always require passed location to simplify get_events_transient_key
The reason that `get_events_transient_key()` could either accept a passed `$location` or use `$this->user_location` was because `$this->user_location` is outdated when the location changes, until the class is re-initialized on the next page load. Because of that, if `$this->user_location` were always used, then the key would be different when cached events were stored than when they were retrieved.

To avoid that behavior, `get_transient_event_key()` could optionally accept the passed `$location` and use that instead. That was a code smell, though, because this function was aware of, and interacting with the controller's state, rather than being a pure model. To make it worse, that fact wasn't documented, making debugging more difficult and time consuming.

It's simpler and more architectually-consistent to just have `get_transient_event_key()` always do the same thing, and let the calling function decide which data to pass based on it's higher-level knowledge of the application state.