Render API errors with a template to avoid XSS and improve UX
The previous method was vulnerable to XSS if an attacker could control the response to the HTTP request (e.g., a man-in-the-middle attack). It was also poor UX, because it would override the entire widget, the user would have to reload the page to get the form back and try another search.

It would probably be safe to use `text()` instead of an Underscore.js template in this case, but that's generally a bad practice. `text()` can still be exploited in some situations, so using it requires the developer to have a deeper understanding of XSS than most. It's much simpler and less error-prone to use an Underscore.js template instead.

The XSS issue is probably just theoretical, since the request is made with HTTPS, but it's best to err on the side of caution. Removing it will also save us from having to reply to lots of reports from security researchers who scan the code with automated tools and see that there's a theoretical vulnerability.