Suggest trying an airport code if the city can't be found
The API will also match airport codes, which is much more likely to succeed than a city search. This is only suggested if the user is having trouble, though, because we want to keep the instructions as simple as possible, and because many people won't know their closest airport code.