Layout style improvements for small screens
This attempts to address problems with the layout interaction between
the event icon, event info, and event date/time that arise on narrow
screens. It uses `display: table-*` style declarations, which are
[cross-browser compatible](http://caniuse.com/#feat=css-table). In my
testing, this solution works for the 350, 800, and 1002 px widths
mentioned in #28.

At the narrowest width, when the event icon is hidden, this also changes
the layout so that the date/time block is below the info block, and the
content is all aligned left.