Escape the city name
There's no valid use of HTML in this variable, and outputing it without escaping creates an XSS vulnerability if an attacker can control the response to the HTTP request (e.g., a man-in-the-middle attack).

That's probably just theoretical, since the request is made with HTTPS, but it's best to err on the side of caution. Removing it will also save us from having to reply to lots of reports from security researchers who scan the code with automated tools and see that there's a theoretical vulnerability.