Make the Edit Location button a true toggle
Before, the button revealed the Edit Location form when clicked, but it
didn't hide it if clicked again. The Cancel button needed to be clicked
instead. This had the potential to cause confusion for screen reader
users.

With this change, the button is a true show/hide toggle, and
the Cancel button has been removed, since it is now redundant.

Refs #36