Handle various error conditions gracefully
Previously, if a string was returned, the error handler would generate a runtime JavaScript error because it was trying to treat a string as if it were an object.

Also, if an object was returned and the object.error property was a function, the handler would just pass it to the template as if it were a string, and the function's raw code would be displayed.