Refactor Events Widget rendering
This changes how the widget UI renders so that only dynamic areas
re-render when the location changes. The toggle button and the search
form now do not re-render, which is a better experience for a11y (and
probably in general).

This also improves the loading spinner position while the location
search is happening.

Refs #43 Refs #40