Log API responses instead of adding to AJAX response, even after merge
Debugging info will still be helpful even after the merge, for cases where users don't get the location they searched for, and we need to know the $request_url to debug it.

The PHP log will be easier to work with in many cases, since it's permenant. It's also a more appropriate method for Core.