Further modularize temmplates to avoid unnecessary re-rendering
This was started in 7350e6e. This commit expands on the work that was done there, so that now all templates are broken into their smallest logical unit.

This was difficult to break into smaller commits, since the templates were previously interconnected to some extent, and because the nature of them touches a lot of things.

At a high level, this does the following:

* Modularizes the templates so that only the element that need to be re-rendered has to be touched
* Moves rendering logic from the view to the controller, where it belongs
* Creates a named function for toggleLocationForm (which was previously a closure) so that it can be reused. This also makes `init()` more readable.
* Ties the pencil icon to the templates were it makes sense, rather than always displaying it

See #43
Fixes #47