Add `speak` messages for when the widget UI updates
This moves the existing `wp.speak` calls from the `getEvents` function
to the `renderEventsTemplate` function, and adds an additional call for
when the widget successfully gets a new location and list of events. The
reason the calls were moved is because A) in this case `wp.speak` is the
audio version of rendering, giving queues that the UI has changed, and
B) it's DRYer because we don't have to duplicate the if/then
conditionals structure that determines what happens based on the
contents of the Ajax response.

This also enqueues the `wp-a11y` script so that the `wp.speak`
functionality is available.

Refs #41