Rename variables to add meaning
Every variable stores data, so naming a variable `data` is essentially meaningless, no different than if it were named `t` or `m`.

The `events` variable sometimes contains events, but it can also contain error messages, locations, etc. The variable should be named to reflect its essential nature, which is a response to an AJAX request.