Change route to `nearby_events/me`, use wp-api.js instead of ajax
Using a wp-api.js Backbone model makes it easy to send requests and
receive responses via the REST API. However, wp-api.js will only
generate a model for a route of there's regex at the end of it, or, as
an exception for `users`, if it ends with "me". Since the nearby events
are related to the current user, "me" seems apropos for this use case
as well.